App web/requirements.txt
absl-py==0.9.0
as==0.1
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.9.1
dash-bootstrap-components==0.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Compress==1.4.0
future==0.18.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
np==1.0.2
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
Pillow==7.0.0
plotly==4.5.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
